backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3
groq==0.9.0
httpx==0.27.0
python-multipart==0.0.12



